컨트롤러 하나로 병합
