+++ src/main/kotlin/com/sparta/techTree/comment/controller/CommentController.kt
@@ -1,7 +1,9 @@
 package com.sparta.techTree.comment.controller
 
 import com.sparta.techTree.comment.dto.CommentDTO
+import com.sparta.techTree.comment.dto.UpdateCommentRequest
 import com.sparta.techTree.comment.service.CommentService
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
@@ -19,5 +21,24 @@ class CommentController(private val commentService: CommentService) {
     fun createComment(@RequestBody commentDto: CommentDTO): ResponseEntity<CommentDTO> {
         val createdComment = commentService.createComment(commentDto)
         return ResponseEntity.ok(createdComment)
+    }
+    @PutMapping("/{commentId}")
+    fun updateComment(
+        @PathVariable commentId: Long,
+        @RequestHeader userId: Long,
+        @RequestBody updateCommentRequest: UpdateCommentRequest
+    ):  ResponseEntity<CommentDTO>{
+        return ResponseEntity
+            .status(HttpStatus.OK)
+            .body(commentService.updateComment(commentId,userId, updateCommentRequest))
+
+    }
+    @DeleteMapping("/{commentId}")
+    fun deleteComment(@PathVariable commentId: Long,@RequestHeader userId: Long,):ResponseEntity<Unit>
+    {
+        commentService.deleteComment(commentId, userId)
+        return ResponseEntity
+            .status(HttpStatus.NO_CONTENT)
+            .build()
     }
 }
